requirements.txt
pymysql==1.0.2
pyyaml==6.0
requests==2.26.0